+++ dart/lib/src/protocol/sentry_user.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import '../../sentry.dart';
+import 'access_aware_map.dart';
 
 /// Describes the current user associated with the application, such as the
 /// currently signed in user.
@@ -44,6 +45,7 @@ class SentryUser {
     Map<String, dynamic>? data,
     @Deprecated('Will be removed in v8. Use [data] instead')
     Map<String, dynamic>? extras,
+    this.unknown,
   })  : assert(
           id != null ||
               username != null ||
@@ -92,8 +94,13 @@ class SentryUser {
   /// Human readable name of the user.
   final String? name;
 
+  @internal
+  final Map<String, dynamic>? unknown;
+
   /// Deserializes a [SentryUser] from JSON [Map].
-  factory SentryUser.fromJson(Map<String, dynamic> json) {
+  factory SentryUser.fromJson(Map<String, dynamic> jsonData) {
+    final json = AccessAwareMap(jsonData);
+
     var extras = json['extras'];
     if (extras != null) {
       extras = Map<String, dynamic>.from(extras);
@@ -120,13 +127,15 @@ class SentryUser {
       name: json['name'],
       // ignore: deprecated_member_use_from_same_package
       extras: extras,
+      unknown: json.notAccessed(),
     );
   }
 
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
     final geoJson = geo?.toJson();
-    return <String, dynamic>{
+    return {
+      ...?unknown,
       if (id != null) 'id': id,
       if (username != null) 'username': username,
       if (email != null) 'email': email,
@@ -165,6 +174,7 @@ class SentryUser {
       extras: extras ?? this.extras,
       geo: geo ?? this.geo,
       name: name ?? this.name,
+      unknown: unknown,
     );
   }
 }