+++ dart/lib/src/sentry_trace_context_header.dart
@@ -1,6 +1,7 @@
 import 'package:meta/meta.dart';
 
 import 'protocol/sentry_id.dart';
+import 'protocol/access_aware_map.dart';
 import 'sentry_baggage.dart';
 import 'sentry_options.dart';
 
@@ -15,6 +16,7 @@ class SentryTraceContextHeader {
     this.transaction,
     this.sampleRate,
     this.sampled,
+    this.unknown,
     this.replayId,
   });
 
@@ -31,10 +33,14 @@ class SentryTraceContextHeader {
   final String? sampled;
 
   @internal
+  final Map<String, dynamic>? unknown;
+
+  @internal
   SentryId? replayId;
 
   /// Deserializes a [SentryTraceContextHeader] from JSON [Map].
-  factory SentryTraceContextHeader.fromJson(Map<String, dynamic> json) {
+  factory SentryTraceContextHeader.fromJson(Map<String, dynamic> data) {
+    final json = AccessAwareMap(data);
     return SentryTraceContextHeader(
       SentryId.fromId(json['trace_id']),
       json['public_key'],
@@ -47,12 +53,14 @@ class SentryTraceContextHeader {
       sampled: json['sampled'],
       replayId:
           json['replay_id'] == null ? null : SentryId.fromId(json['replay_id']),
+      unknown: json.notAccessed(),
     );
   }
 
   /// Produces a [Map] that can be serialized to JSON.
   Map<String, dynamic> toJson() {
     return {
+      ...?unknown,
       'trace_id': traceId.toString(),
       'public_key': publicKey,
       if (release != null) 'release': release,