+++ dart/lib/src/transport/http_transport.dart
@@ -11,6 +11,8 @@ import '../sentry_options.dart';
 import '../sentry_envelope.dart';
 import 'transport.dart';
 import 'rate_limiter.dart';
+import '../http_client/client_provider.dart'
+    if (dart.library.io) '../http_client/io_client_provider.dart';
 
 /// A transport is in charge of sending the event to the Sentry server.
 class HttpTransport implements Transport {
@@ -22,9 +24,8 @@ class HttpTransport implements Transport {
 
   factory HttpTransport(SentryOptions options, RateLimiter rateLimiter) {
     if (options.httpClient is NoOpClient) {
-      options.httpClient = Client();
+      options.httpClient = getClientProvider().getClient(options);
     }
-
     return HttpTransport._(options, rateLimiter);
   }