+++ flutter/lib/src/native/sentry_native_channel.dart
@@ -66,6 +66,7 @@ class SentryNativeChannel
       'readTimeoutMillis': options.readTimeout.inMilliseconds,
       'appHangTimeoutIntervalMillis':
           options.appHangTimeoutInterval.inMilliseconds,
+      if (options.proxy != null) 'proxy': options.proxy?.toJson(),
       'replay': <String, dynamic>{
         'sessionSampleRate': options.experimental.replay.sessionSampleRate,
         'errorSampleRate': options.experimental.replay.errorSampleRate,